+++ docs/CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 - Add processes ``alignment-star-new`` and
   ``alignment-star-index-new`` that use STAR version
   2.7.10b
+- Add process ``star-quantification``
 
 Changed
 -------